Fix a compilation warning in `libelf`.
git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@2252 95820547-d848-0410-985e-9ae8fe0fa350

+++ libelf/elf_begin.c
@@ -200,7 +200,7 @@ _libelf_open_object(int fd, Elf_Cmd c)
 				return (NULL);
 			}
 
-			if (read(fd, m, fsize) != fsize) {
+			if (read(fd, m, fsize) != (ssize_t) fsize) {
 				LIBELF_SET_ERROR(IO, errno);
 				free(m);
 				return (NULL);
